svg/painting/parsing/stroke-miterlimit-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing stroke-miterlimit with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeMiterlimitProperty" rel="help"/><h:meta content="stroke-miterlimit supports only the grammar 'number'" name="assert"/><h:meta content="A negative value for stroke-miterlimit must be treated as an illegal value." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("stroke-miterlimit", "1.");
test_invalid_value("stroke-miterlimit", "2 3");
test_invalid_value("stroke-miterlimit", "-4");

  ]]></script></svg>
